Create min_product_subset_array.cpp
Input : a[] = { -1, -1, -2, 4, 3 }
Output : -24
Explanation : Minimum product will be ( -2 * -1 * -1 * 4 * 3 ) = -24

If there are even number of negative numbers and no zeros, the result is the product of all except the largest valued negative number.
If there are an odd number of negative numbers and no zeros, the result is simply the product of all.
If there are zeros and positive, no negative, the result is 0. The exceptional case is when there is no negative number and all other elements positive then our result should be the first minimum positive number.